predict scent tracking logic (move to shared)
